requirements.txt
fpdf2
markdown
python-docx
fastapi
uvicorn
python-multipart
uharfbuzz
pypdf
